Tratamento de exceptions

+++ src/main/java/org/serratec/alterstatusapi/service/UsuarioServiceImpl.java
@@ -12,6 +12,7 @@ import java.util.Optional;
 import javax.transaction.Transactional;
 
 import org.serratec.alterstatusapi.dto.UsuarioDTORequest;
+import org.serratec.alterstatusapi.exception.ResourceBadRequestException;
 import org.serratec.alterstatusapi.exception.ResourceNotFoundException;
 import org.serratec.alterstatusapi.mapper.UsuarioMapper;
 import org.serratec.alterstatusapi.model.Cargo;
@@ -90,7 +91,13 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public Usuario obterPorId(Long id) throws ResourceNotFoundException {
+		
+		if(id == null) {
+			throw new ResourceBadRequestException("Nenhum id foi definido");
+		}
+		
 		Optional<Usuario> cliente = clienteRepository.findById(id);
+		
 		if (cliente.isEmpty()) {
 			throw new ResourceNotFoundException("Não existe cliente com esse Id.");
 		}
@@ -100,11 +107,18 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public List<Usuario> obterPorUsername(String username) {
+		
+		if(username == null) {
+			throw new ResourceBadRequestException("Username não foi definido");
+		}
+		
 		List<Usuario> usuario = clienteRepository.findByUsernameContainingIgnoreCase(username);
+		
 		if (usuario.isEmpty()) {
 			throw new ResourceNotFoundException("Não foi localizado nenhum usuário com o username: " + username);
 
 		}
+		
 		return usuario;
 	}
 
@@ -129,6 +143,17 @@ public class UsuarioServiceImpl implements UsuarioService {
 	@Override
 	public Usuario cadastrarArquivo(UsuarioDTORequest dto, MultipartFile multipartFile)
 			throws ResourceNotFoundException, IOException {
+		
+		if(dto.getNome() == "" || dto.getSenha() == "" || dto.getUsername() == "") {
+			throw new ResourceBadRequestException("Um dos campos não foi definido");
+		}
+		
+		Usuario UsuarioVerificar = clienteRepository.findByUsername(dto.getUsername());
+		
+		if(UsuarioVerificar != null) {
+			throw new ResourceBadRequestException("Username ja existe");
+		}
+		
 		Usuario entity = clienteMapper.toEntity(dto);
 
 		entity.setDataCadastro(LocalDate.now());
@@ -151,6 +176,17 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public Usuario cadastrar(UsuarioDTORequest dto) throws ResourceNotFoundException {
+		
+		if(dto.getNome() == "" || dto.getSenha() == "" || dto.getUsername() == "") {
+			throw new ResourceBadRequestException("Um dos campos não foi definido");
+		}
+		
+		Usuario UsuarioVerificar = clienteRepository.findByUsername(dto.getUsername());
+		
+		if(UsuarioVerificar != null) {
+			throw new ResourceBadRequestException("Username ja existe");
+		}
+		
 		Usuario entity = clienteMapper.toEntity(dto);
 
 		entity.setDataCadastro(LocalDate.now());
@@ -170,7 +206,13 @@ public class UsuarioServiceImpl implements UsuarioService {
 	}
 	
 	@Override
-	public Usuario atualizarArquivo(Long id, UsuarioDTORequest dto, MultipartFile multipartFile) throws ResourceNotFoundException, IOException {
+	public Usuario atualizarArquivo(Long id, UsuarioDTORequest dto, MultipartFile multipartFile) 
+			throws ResourceNotFoundException, IOException {
+		
+		if(dto.getNome() == "" || dto.getSenha() == "" || dto.getUsername() == "" || id == null) {
+			throw new ResourceBadRequestException("Um dos campos não foi definido");
+		}
+		
 		Usuario entity = this.obterPorId(id);
 
 		for (Usuario cliente : this.obterTodos()) {
@@ -204,6 +246,11 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public Usuario atualizar(Long id, UsuarioDTORequest dto) throws ResourceNotFoundException {
+		
+		if(dto.getNome() == "" || dto.getSenha() == "" || dto.getUsername() == "" || id == null) {
+			throw new ResourceBadRequestException("Um dos campos não foi definido");
+		}
+		
 		Usuario entity = this.obterPorId(id);
 
 		for (Usuario cliente : this.obterTodos()) {
@@ -233,6 +280,11 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public void deletar(Long id) throws ResourceNotFoundException {
+		
+		if(id == null) {
+			throw new ResourceBadRequestException("Nenhum id definido");
+		}
+		
 		this.obterPorId(id); // verifica se o usuário existe antes de deletar
 
 		imagemService.removerImagem(id);
@@ -243,6 +295,11 @@ public class UsuarioServiceImpl implements UsuarioService {
 	@Transactional
 	@Override
 	public void removerCargo(Long idUsuario) throws ResourceNotFoundException {
+		
+		if(idUsuario == null) {
+			throw new ResourceBadRequestException("Nenhum id definido");
+		}
+		
 		Usuario usuario = this.obterPorId(idUsuario); // verifica se o usuário existe antes de deletar
 
 		usuario.setCargo(null);
@@ -252,8 +309,14 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public Usuario relacionarUsuarioComSquad(Long usuario_id, Long squad_id) {
+		
+		if(usuario_id == null || squad_id == null) {
+			throw new ResourceBadRequestException("Um dos id(s) não foi definido");
+		}
+		
 		Usuario usuario = clienteRepository.findById(usuario_id).get();
 		Squad squad = repositorioSquad.findById(squad_id).get();
+		
 		usuario.setId_squad(squad_id);
 		usuario.setSquad(squad);
 
@@ -262,8 +325,14 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public Usuario relacionarUsuarioComCargo(Long usuario_id, Long cargo_id) {
+		
+		if(usuario_id == null || cargo_id == null) {
+			throw new ResourceBadRequestException("Um dos id(s) não foi definido");
+		}
+		
 		Usuario usuario = clienteRepository.findById(usuario_id).get();
 		Cargo cargo = repositorioCargo.findById(cargo_id).get();
+		
 		usuario.setId_cargo(cargo_id);
 		usuario.setCargo(cargo);
 
@@ -272,6 +341,11 @@ public class UsuarioServiceImpl implements UsuarioService {
 
 	@Override
 	public ResponseEntity<Optional<Usuario>> atualizarEspecifico(Long id, Map<Object, Object> campos) {
+		
+		if(id == null || campos == null) {
+			throw new ResourceBadRequestException("Um dos campos não foi definido");
+		}
+		
 		Optional<Usuario> UsuarioAtualizado = clienteRepository.findById(id);
 		// Map key is field name, v is value
 		campos.forEach((chave, valor) -> {
